Create a 'Close' Button to Exit the Survey

This function creates a 'Close' button that, when clicked, will close the current browser tab or window. The button can be activated by clicking or by pressing the Enter key.

Usage

sd_close(label = "Exit Survey")

Arguments

label

Character string. The label of the 'Close' button. Defaults to "Exit Survey".

Value

A Shiny action button UI element with associated JavaScript for closing the page and Enter key functionality.

Details

The function generates a Shiny action button that, when clicked or when the Enter key is pressed, will attempt to close the current browser tab or window. Note that for security reasons, some browsers may not allow JavaScript to close windows that were not opened by JavaScript. In such cases, the button will prompt the user to close the tab manually.

Examples

sd_close()
#> <div style="margin-top: 0.5rem; margin-bottom: 0.5rem;">
#>   <button id="close-survey-button" type="button" class="btn btn-default action-button" style="display: block; margin: auto;" onclick="window.close(); if (!window.closed) { alert(&#39;Please close this tab manually to exit the survey.&#39;); }">Exit Survey</button>
#> </div>
#> <script>
#>     $(document).ready(function() {
#>         var buttonId = 'close-survey-button';
#>         $(document).on('keydown', function(event) {
#>             if (event.key === 'Enter' && !event.repeat) {
#>                 var $button = $('#' + buttonId);
#>                 if ($button.is(':visible')) {
#>                     $button.click();
#>                     event.preventDefault();
#>                 }
#>             }
#>         });
#>     });
#>     </script>
sd_close("Exit Survey")
#> <div style="margin-top: 0.5rem; margin-bottom: 0.5rem;">
#>   <button id="close-survey-button" type="button" class="btn btn-default action-button" style="display: block; margin: auto;" onclick="window.close(); if (!window.closed) { alert(&#39;Please close this tab manually to exit the survey.&#39;); }">Exit Survey</button>
#> </div>
#> <script>
#>     $(document).ready(function() {
#>         var buttonId = 'close-survey-button';
#>         $(document).on('keydown', function(event) {
#>             if (event.key === 'Enter' && !event.repeat) {
#>                 var $button = $('#' + buttonId);
#>                 if ($button.is(':visible')) {
#>                     $button.click();
#>                     event.preventDefault();
#>                 }
#>             }
#>         });
#>     });
#>     </script>