Create a Survey Template

This function creates a survey template in a specified directory. It can use different templates, with "simple" being the default. The function prompts the user to confirm the use of the current working directory if no path is specified.

sd_create_survey(path = getwd(), template = "simple")

Arguments

path

A character string specifying the directory in which to create the survey template. Defaults to the current working directory.

template

A character string specifying the survey template to use. Defaults to "simple".

Value

A message indicating the successful creation of the survey template.

Examples

if (FALSE) { # \dontrun{
sd_create_survey()
sd_create_survey(path = "path/to/package", template = "simple")
} # }