Configuration Function for surveydown Surveys

This function sets up the configuration for a surveydown survey, including page and question structures, conditional display settings, and navigation options.

Usage

sd_config(
  skip_if = NULL,
  skip_if_custom = NULL,
  show_if = NULL,
  show_if_custom = NULL,
  required_questions = NULL,
  all_questions_required = FALSE,
  start_page = NULL,
  show_all_pages = FALSE,
  admin_page = FALSE
)

Arguments

skip_if

A list of conditions under which certain pages should be skipped. Defaults to NULL.

skip_if_custom

A custom function to handle conditions under which certain pages should be skipped. Defaults to NULL.

show_if

A list of conditions under which certain pages should be shown. Defaults to NULL.

show_if_custom

A custom function to handle conditions under which certain pages should be shown. Defaults to NULL.

required_questions

Vector of character strings. The IDs of questions that must be answered before the respondent can continue in the survey or survey can be submitted. Defaults to NULL.

all_questions_required

Logical. If TRUE, all questions in the survey will be required. This overrides the required_questions parameter. Defaults to FALSE.

start_page

Character string. The ID of the page to start on. Defaults to NULL.

show_all_pages

Logical. Whether to show all pages initially. Defaults to FALSE.

admin_page

Logical. Whether to include an admin page for viewing and downloading survey data. Defaults to FALSE.

Value

A list containing the configuration settings for the survey, including page and question structures, conditional display settings, navigation options, and admin page settings.

Details

The function retrieves the survey metadata, checks the validity of the conditional display settings, and ensures that the specified start page (if any) exists. It then stores these settings in a configuration list. If admin_page is set to TRUE, an admin page will be included in the survey. This page allows viewing and downloading of survey data upon entering the correct survey password (set using sd_set_password()).

If all_questions_required is set to TRUE, it will override the required_questions parameter and set all questions in the survey as required.

Examples

# Examples here