Connect to a supabase Database

This function establishes a connection to a supabase database using the provided connection details.

Usage

sd_database(
  host = NULL,
  dbname = NULL,
  port = NULL,
  user = NULL,
  table_name = NULL,
  password = Sys.getenv("SURVEYDOWN_PASSWORD"),
  gssencmode = "prefer",
  pause = FALSE
)

Arguments

host

Character string. The host address of the supabase database.

dbname

Character string. The name of the supabase database.

port

Integer. The port number for the supabase database connection.

user

Character string. The username for the supabase database connection.

table_name

Character string. The name of the table to interact with in the supabase database.

password

Character string. The password for the supabase database connection. Defaults to the value of the SURVEYDOWN_PASSWORD environment variable.

gssencmode

Character string. The GSS encryption mode for the database connection. Defaults to "prefer".

pause

Logical. If TRUE, data will be saved to a local CSV file instead of the database. Defaults to FALSE.

Value

A list containing the database connection object (db) and the table name (table_name), or NULL if in pause mode.

Details

The function checks for the presence of all required parameters and attempts to establish a connection to the supabase database. If successful, it returns a list containing the database connection object and the table name. The user must have created the specified table in supabase beforehand. If pause mode is enabled, the function returns NULL and data will be saved to a local CSV file. The password is obtained from the SURVEYDOWN_PASSWORD environment variable by default, but can be overridden by explicitly passing a value.

Note

The user must create their own table inside supabase in order to make additions.

Examples

if (FALSE) { # \dontrun{
  # Assuming SURVEYDOWN_PASSWORD is set in .Renviron
  db_connection <- sd_database(
    host       = "aws-0-us-west-1.pooler.supabase.com",
    dbname     = "postgres",
    port       = "6---",
    user       = "postgres.k----------i",
    table_name = "your-table-name",
    pause      = FALSE
  )

  # Explicitly providing the password
  db_connection <- sd_database(
    host       = "aws-0-us-west-1.pooler.supabase.com",
    dbname     = "postgres",
    port       = "6---",
    user       = "postgres.k----------i",
    table_name = "your-table-name",
    password   = "your-password",
    pause      = FALSE
  )
} # }