Server Logic for a surveydown survey

This function defines the server-side logic for a Shiny application used in surveydown. It handles various operations such as conditional display, progress tracking, page navigation, and database updates for survey responses.

Usage

sd_server(
  db = NULL,
  use_html = FALSE,
  required_questions = NULL,
  all_questions_required = FALSE,
  start_page = NULL,
  admin_page = FALSE
)

Arguments

db

A list containing database connection information created using sd_database function. Defaults to NULL.

use_html

Logical. By default, the "survey.qmd" file will be rendered when the app launches, which can be slow. Users can render it first into a html file and set use_html = TRUE to use the pre-rendered file, which is faster when the app loads. Defaults to FALSE.

required_questions

Vector of character strings. The IDs of questions that must be answered. Defaults to NULL.

all_questions_required

Logical. If TRUE, all questions in the survey will be required. Defaults to FALSE.

start_page

Character string. The ID of the page to start on. Defaults to NULL.

admin_page

Logical. Whether to include an admin page for viewing and downloading survey data. Defaults to FALSE.

Value

This function does not return a value; it sets up the server-side logic for the Shiny application.

Details

The function performs the following tasks:

Initializes variables and reactive values.

Implements conditional display logic for questions.

Tracks answered questions and updates the progress bar.

Handles page navigation and skip logic.

Manages required questions.

Performs database operations or saves to a local CSV file in preview mode.

Sets up admin functionality if enabled in the configuration.

Progress Bar

The progress bar is updated based on the last answered question. It will jump to the percentage corresponding to the last answered question and will never decrease, even if earlier questions are answered later. The progress is calculated as the ratio of the last answered question's index to the total number of questions.

Database Operations

If db is provided, the function will update the database with survey responses. If db is NULL (ignore mode), responses will be saved to a local CSV file.

See also

Examples

if (FALSE) { # \dontrun{
  library(surveydown)
  db <- sd_database()

  shinyApp(
    ui = sd_ui(),
    server = function(input, output, session) {
      sd_server(db = db)
    }
  )
} # }